requirements.txt
alembic==1.4.2
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cachelib==0.4.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
convertdate==2.3.0
decorator==4.4.2
distlib==0.3.0
filelock==3.4.0
Flask==1.1.2
Flask-Dance==3.1.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
flask-talisman==1.0.0
Flask-WTF==1.0.0
google-api-core==2.6.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.55.0
idna==2.10
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
Mako==1.1.6
MarkupSafe==1.1.1
mysql-client==0.0.1
mysqlclient==1.4.6
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.19.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
PyMeeus==0.5.11
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.2
URLObject==2.4.3
virtualenv==20.0.20
wcwidth==0.1.9
Werkzeug==1.0.1
whl==0.0.1
WTForms==3.0.1
zipp==3.1.0
